+++ athena/src/main/java/goorm/athena/domain/userCoupon/service/test/UserCouponCommandServiceV4_5.java
@@ -1,137 +1,116 @@
 package goorm.athena.domain.userCoupon.service.test;
 
-import goorm.athena.domain.coupon.entity.Coupon;
-import goorm.athena.domain.coupon.service.CouponQueryService;
-import goorm.athena.domain.user.entity.User;
-import goorm.athena.domain.user.service.UserQueryService;
 import goorm.athena.domain.userCoupon.dto.req.UserCouponIssueRequest;
-import goorm.athena.domain.userCoupon.dto.res.UserCouponIssueResponse;
-import goorm.athena.domain.userCoupon.entity.UserCoupon;
-import goorm.athena.domain.userCoupon.mapper.UserCouponMapper;
-import goorm.athena.domain.userCoupon.repository.UserCouponRepository;
+import goorm.athena.domain.userCoupon.event.UserCouponIssueEvent;
+import goorm.athena.domain.coupon.event.CouponSyncTriggerEvent;
 import goorm.athena.global.exception.CustomException;
 import goorm.athena.global.exception.ErrorCode;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.redisson.api.RScript;
 import org.redisson.api.RSet;
 import org.redisson.api.RedissonClient;
 import org.redisson.client.codec.StringCodec;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.Arrays;
 import java.util.List;
 
 /*
-    @Async 기반 비동기 처리로 메인에서 Redis 처리 후, 서브 스레드에서 DB 처리 진행함
-    비동기 처리로 병렬화가 가능하나 내부 스레드 기반으로 확장성에 한계가 존재함
-    또한, Redis Set을 사용하여 Lock을 사용하지 않음으로써 Lock으로 발생하는 부하를 줄임
+    Redis Lua 스크립트 내 재고 관리 및 플래그 처리를 원자적으로 계산하도록 함
+    DB까지 가지않음으로써 빠른 성능 개선이 구현됨
+    단, Lua Script는 Redis를 이용한 것이기에 Redis 부하가 발생할 시 성능 저하 우려
  */
+
 @Service
 @RequiredArgsConstructor
 public class UserCouponCommandServiceV4_5 {
     private final RedissonClient redissonClient;
-    private final UserQueryService userQueryService;
-    private final CouponQueryService couponQueryService;
-    private final UserCouponRepository userCouponRepository;
-    private final UserCouponMapper userCouponMapper;
     private final ApplicationEventPublisher eventPublisher;
 
     private static final String LUA_SCRIPT = """
         local total = tonumber(redis.call('GET', KEYS[1]))
         local used = tonumber(redis.call('GET', KEYS[2])) or 0
-
+        local flagKey = KEYS[3]
+    
         if not total then
             return -1  -- 쿠폰 없음
         end
-
+    
         if used >= total then
             return 0  -- 품절
         end
-
+    
         redis.call('INCR', KEYS[2])
-        return 1  -- 성공
+        used = used + 1
+    
+        if used >= total then
+            local setFlag = redis.call('SETNX', flagKey, '1')
+            if setFlag == 1 then
+                redis.call('EXPIRE', flagKey, 60)  -- 플래그 TTL 설정
+                return 2  -- 마지막 쿠폰 발급 + 플래그 세팅 완료
+            else
+                return 2  -- 플래그 이미 세팅됨, 재고 소진 상태
+            end
+        end
+    
+        return 1  -- 정상 발급
     """;
 
-    // Redis 재고 감소 처리 (트랜잭션 외부)
-    public void checkAndDecreaseRedisStock(Long couponId) {
-        String totalKey = "coupon_total_" + couponId;
-        String usedKey = "coupon_used_" + couponId;
-
-        RScript script = redissonClient.getScript();
-        List<Object> keys = Arrays.asList(totalKey, usedKey);
-
-        Long result = script.eval(
-                RScript.Mode.READ_WRITE,
-                LUA_SCRIPT,
-                RScript.ReturnType.INTEGER,
-                keys
-        );
-
-        if (result == null || result == -1) {
-            throw new CustomException(ErrorCode.COUPON_NOT_FOUND);
-        } else if (result == 0) {
-            throw new CustomException(ErrorCode.COUPON_OUT_STOCK);
-        }
-    }
-
-    // DB 트랜잭션 내 저장 및 중복 체크
-    @Transactional
-    public UserCouponIssueResponse saveCouponIssue(Long userId, Long couponId) {
-        User user = userQueryService.getUser(userId);
-        Coupon coupon = couponQueryService.getCoupon(couponId);
-
-        try {
-            UserCoupon userCoupon = UserCoupon.create(user, coupon);
-            userCouponRepository.save(userCoupon);
-            return userCouponMapper.toCreateResponse(userCoupon);
-        } catch (DataIntegrityViolationException e) {
-            // 중복 발급 예외 처리
-            throw new CustomException(ErrorCode.ALREADY_ISSUED_COUPON);
-        }
-    }
-
-    // Redis Set 중복 체크 & 등록 메서드
-    private boolean addUserToIssuedSet(Long couponId, Long userId) {
-        String issuedSetKey = "issued_users_" + couponId;
-        RSet<String> issuedSet = redissonClient.getSet(issuedSetKey, StringCodec.INSTANCE);
-        return issuedSet.add(String.valueOf(userId));
-    }
-
-    private void removeUserFromIssuedSet(Long couponId, Long userId) {
-        String issuedSetKey = "issued_users_" + couponId;
-        RSet<String> issuedSet = redissonClient.getSet(issuedSetKey, StringCodec.INSTANCE);
-        issuedSet.remove(String.valueOf(userId));
-    }
-
-    // 전체 쿠폰 발급 플로우 (Redis 재고 + Redis Set 중복 체크 통합)
-    public UserCouponIssueResponse issueCoupon(Long userId, UserCouponIssueRequest request) {
+    public void issueCoupon(Long userId, UserCouponIssueRequest request) {
         Long couponId = request.couponId();
 
-        // 1) Redis Set 중복 체크 및 등록 (중복 시 예외)
+        // 1. 중복 체크 & 등록
         boolean added = addUserToIssuedSet(couponId, userId);
         if (!added) {
             throw new CustomException(ErrorCode.ALREADY_ISSUED_COUPON);
         }
 
         try {
-            // 2) Redis 재고 감소 체크
-            checkAndDecreaseRedisStock(couponId);
+            // 2. Redis 재고 체크 및 감소 (플래그 세팅까지 Lua에서 처리)
+            int result = checkAndDecreaseRedisStock(couponId);
 
-            // 3) DB 저장 (중복 예외 시 롤백)
-            UserCouponIssueResponse response = saveCouponIssue(userId, couponId);
+            // 3. 발급 성공 시 이벤트 발행
+            eventPublisher.publishEvent(new UserCouponIssueEvent(userId, couponId, result));
 
-            // 4) 이벤트 발행
-            return response;
-        } catch (CustomException e) {
-            // Redis Set에서 userId 제거(롤백)
-            removeUserFromIssuedSet(couponId, userId);
-            throw e;
+            // 4. 재고 소진 플래그가 Lua 스크립트에서 세팅되었을 경우 동기화 이벤트 발행
+            if (result == 2) {
+                eventPublisher.publishEvent(new CouponSyncTriggerEvent(couponId));
+            }
         } catch (RuntimeException e) {
-            removeUserFromIssuedSet(couponId, userId);
+            removeUserFromIssuedSet(couponId, userId); // 롤백
             throw e;
         }
+    }
+
+    private int checkAndDecreaseRedisStock(Long couponId) {
+        String totalKey = "coupon_total_" + couponId;
+        String usedKey = "coupon_used_" + couponId;
+        String flagKey = "coupon_sync_triggered_" + couponId;
+
+        List<Object> keys = Arrays.asList(totalKey, usedKey, flagKey);
+        RScript script = redissonClient.getScript();
+
+        Long result = script.eval(RScript.Mode.READ_WRITE, LUA_SCRIPT, RScript.ReturnType.INTEGER, keys);
+
+        if (result == null || result == -1) {
+            throw new CustomException(ErrorCode.COUPON_NOT_FOUND);
+        } else if (result == 0) {
+            throw new CustomException(ErrorCode.COUPON_OUT_STOCK);
+        }
+
+        return result.intValue();
+    }
+
+    private boolean addUserToIssuedSet(Long couponId, Long userId) {
+        String key = "issued_users_" + couponId;
+        RSet<String> issuedSet = redissonClient.getSet(key, StringCodec.INSTANCE);
+        return issuedSet.add(String.valueOf(userId));
+    }
+
+    private void removeUserFromIssuedSet(Long couponId, Long userId) {
+        String key = "issued_users_" + couponId;
+        RSet<String> issuedSet = redissonClient.getSet(key, StringCodec.INSTANCE);
+        issuedSet.remove(String.valueOf(userId));
     }
 }
